+++ CHANGES.md
@@ -22,12 +22,17 @@ To be released.
 
  -  Added `Router.trailingSlashInsensitive` property.
 
- -  Implemented HTTP Message Signatures ([RFC 9421]).  [[#208]]
+ -  Implemented HTTP Message Signatures ([RFC 9421]) with double-knocking.
+    Currently, it only works with RSA-PKCS#1-v1.5.  [[#208]]
 
      -  Added `HttpMessageSignaturesSpec` type.
      -  Added `SignRequestOptions.spec` option.
      -  Added `SignRequestOptions.currentTime` option.
      -  Added `VerifyRequestOptions.spec` option.
+     -  Added `GetAuthenticatedDocumentLoaderOptions.specDeterminer` option.
+     -  Added `GetAuthenticatedDocumentLoaderOptions.traceProvider` option.
+     -  Added `HttpMessageSignaturesSpecDeterminer` interface.
+     -  Added `--first-knock` option to `fedify lookup` command.
 
 [RFC 9421]: https://www.rfc-editor.org/rfc/rfc9421
 [#208]: https://github.com/fedify-dev/fedify/issues/208